add require log for multicalls
