+++ packages/fuels/tests/contracts.rs
@@ -644,102 +644,3 @@ async fn contract_call_futures_implement_send() -> Result<(), Error> {
     });
     Ok(())
 }
-
-fn assert_is_revert_containing_msg(msg: &str, error: Error) {
-    assert!(matches!(error, Error::RevertTransactionError(..)));
-    if let Error::RevertTransactionError(error_message, _) = error {
-        assert!(error_message.contains(msg));
-    }
-}
-
-#[tokio::test]
-async fn test_require() -> Result<(), Error> {
-    setup_contract_test!(
-        contract_instance,
-        wallet,
-        "packages/fuels/tests/contracts/require"
-    );
-
-    let contract_methods = contract_instance.methods();
-    {
-        let error = contract_methods
-            .require_primitive()
-            .call()
-            .await
-            .expect_err("Should return a revert error");
-        dbg!(&error);
-        assert_is_revert_containing_msg("42", error);
-    }
-    {
-        let error = contract_methods
-            .require_string()
-            .call()
-            .await
-            .expect_err("Should return a revert error");
-
-        dbg!(&error);
-        assert_is_revert_containing_msg("fuel", error);
-    }
-    {
-        let error = contract_methods
-            .require_custom_generic()
-            .call()
-            .await
-            .expect_err("Should return a revert error");
-
-        dbg!(&error);
-        assert_is_revert_containing_msg("StructDeeplyNestedGeneric", error);
-    }
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_multi_call_require() -> Result<(), Error> {
-    setup_contract_test!(
-        contract_instance,
-        wallet,
-        "packages/fuels/tests/contracts/require"
-    );
-
-    let contract_methods = contract_instance.methods();
-
-    // The output of the error depends on the order of the contract
-    // handlers as the script returns the first revert it finds.
-    {
-        let call_handler_1 = contract_methods.require_string();
-        let call_handler_2 = contract_methods.require_custom_generic();
-
-        let mut multi_call_handler = MultiContractCallHandler::new(wallet.clone());
-
-        multi_call_handler
-            .add_call(call_handler_1)
-            .add_call(call_handler_2);
-
-        let error = multi_call_handler
-            .call::<((), ())>()
-            .await
-            .expect_err("Should return a revert error");
-
-        dbg!(&error);
-        assert_is_revert_containing_msg("fuel", error);
-    }
-    {
-        let call_handler_1 = contract_methods.require_custom_generic();
-        let call_handler_2 = contract_methods.require_string();
-
-        let mut multi_call_handler = MultiContractCallHandler::new(wallet.clone());
-
-        multi_call_handler
-            .add_call(call_handler_1)
-            .add_call(call_handler_2);
-
-        let error = multi_call_handler
-            .call::<((), ())>()
-            .await
-            .expect_err("Should return a revert error");
-
-        dbg!(&error);
-        assert_is_revert_containing_msg("StructDeeplyNestedGeneric", error);
-    }
-    Ok(())
-}
